Move inLifecycleMethod to astUtil
