+++ lib/rules/no-unused-prop-types.js
@@ -13,6 +13,7 @@ const variable = require('../util/variable');
 const annotations = require('../util/annotations');
 const versionUtil = require('../util/version');
 const propsUtil = require('../util/props');
+const astUtil = require('../util/ast');
 
 // ------------------------------------------------------------------------------
 // Constants
@@ -21,7 +22,6 @@ const propsUtil = require('../util/props');
 const DIRECT_PROPS_REGEX = /^props\s*(\.|\[)/;
 const DIRECT_NEXT_PROPS_REGEX = /^nextProps\s*(\.|\[)/;
 const DIRECT_PREV_PROPS_REGEX = /^prevProps\s*(\.|\[)/;
-const LIFE_CYCLE_METHODS = ['constructor', 'componentWillReceiveProps', 'shouldComponentUpdate', 'componentWillUpdate', 'componentDidUpdate'];
 
 // ------------------------------------------------------------------------------
 // Rule Definition
@@ -82,24 +82,6 @@ module.exports = {
       return value;
     }
 
-    /**
-     * Check if we are in a lifecycle method
-     * @return {boolean} true if we are in a lifecycle method, false if not
-     **/
-    function inLifeCycleMethod() {
-      let scope = context.getScope();
-      while (scope) {
-        if (
-          scope.block && scope.block.parent &&
-          scope.block.parent.key &&
-            LIFE_CYCLE_METHODS.indexOf(scope.block.parent.key.name) >= 0
-        ) {
-          return true;
-        }
-        scope = scope.upper;
-      }
-      return false;
-    }
 
     /**
      * Check if the current node is in a setState updater method
@@ -155,7 +137,7 @@ module.exports = {
         || isPropArgumentInSetStateUpdater(node))
       );
       const isStatelessFunctionUsage = node.object.name === 'props';
-      return isClassUsage || isStatelessFunctionUsage || inLifeCycleMethod();
+      return isClassUsage || isStatelessFunctionUsage || astUtil.inLifeCycleMethod(context);
     }
 
     /**
@@ -546,7 +528,7 @@ module.exports = {
       const isDirectSetStateProp = isPropArgumentInSetStateUpdater(node);
       const isInClassComponent = utils.getParentES6Component() || utils.getParentES5Component();
       const isNotInConstructor = !inConstructor(node);
-      const isNotInLifeCycleMethod = !inLifeCycleMethod();
+      const isNotInLifeCycleMethod = !astUtil.inLifeCycleMethod(context);
       const isNotInSetStateUpdater = !inSetStateUpdater();
       if ((isDirectProp || isDirectNextProp || isDirectPrevProp || isDirectSetStateProp)
         && isInClassComponent
@@ -636,7 +618,7 @@ module.exports = {
             // let {firstname} = props
             const genericDestructuring = isPropAttributeName(node) && (
               utils.getParentStatelessComponent() ||
-              inLifeCycleMethod()
+              astUtil.inLifeCycleMethod(context)
             );
 
             if (thisDestructuring) {
@@ -958,7 +940,7 @@ module.exports = {
         // let {firstname} = props
         const statelessDestructuring = destructuring && isPropAttributeName(node) && (
           utils.getParentStatelessComponent() ||
-          inLifeCycleMethod()
+          astUtil.inLifeCycleMethod(context)
         );
 
         if (!thisDestructuring && !statelessDestructuring) {
@@ -1024,7 +1006,7 @@ module.exports = {
       ObjectPattern: function(node) {
         // If the object pattern is a destructured props object in a lifecycle
         // method -- mark it for used props.
-        if (inLifeCycleMethod()) {
+        if (astUtil.inLifeCycleMethod(context)) {
           node.properties.forEach((property, i) => {
             if (i === 0) {
               markPropTypesAsUsed(node.parent);
