+++ lib/rules/prop-types.js
@@ -13,6 +13,7 @@ const variable = require('../util/variable');
 const annotations = require('../util/annotations');
 const versionUtil = require('../util/version');
 const propsUtil = require('../util/props');
+const astUtil = require('../util/ast');
 
 // ------------------------------------------------------------------------------
 // Constants
@@ -20,7 +21,6 @@ const propsUtil = require('../util/props');
 
 const PROPS_REGEX = /^(props|nextProps)$/;
 const DIRECT_PROPS_REGEX = /^(props|nextProps)\s*(\.|\[)/;
-const LIFE_CYCLE_METHODS = ['constructor', 'componentWillReceiveProps', 'shouldComponentUpdate', 'componentWillUpdate', 'componentDidUpdate'];
 
 // ------------------------------------------------------------------------------
 // Rule Definition
@@ -88,25 +88,6 @@ module.exports = {
     }
 
     /**
-     * Check if we are in a lifecycle method
-     * @return {boolean} true if we are in a lifecycle method, false if not
-     **/
-    function inLifeCycleMethod() {
-      let scope = context.getScope();
-      while (scope) {
-        if (
-          scope.block && scope.block.parent &&
-          scope.block.parent.key &&
-            LIFE_CYCLE_METHODS.indexOf(scope.block.parent.key.name) >= 0
-        ) {
-          return true;
-        }
-        scope = scope.upper;
-      }
-      return false;
-    }
-
-    /**
     * Checks if a prop is being assigned a value props.bar = 'bar'
     * @param {ASTNode} node The AST node being checked.
     * @returns {Boolean}
@@ -131,7 +112,7 @@ module.exports = {
         node.object.type === 'ThisExpression' && node.property.name === 'props'
       );
       const isStatelessFunctionUsage = node.object.name === 'props' && !isAssignmentToProp(node);
-      const isNextPropsUsage = node.object.name === 'nextProps' && inLifeCycleMethod();
+      const isNextPropsUsage = node.object.name === 'nextProps' && astUtil.inLifeCycleMethod(context);
       return isClassUsage || isStatelessFunctionUsage || isNextPropsUsage;
     }
 
@@ -521,7 +502,7 @@ module.exports = {
     function getPropertyName(node) {
       const isDirectProp = DIRECT_PROPS_REGEX.test(sourceCode.getText(node));
       const isInClassComponent = utils.getParentES6Component() || utils.getParentES5Component();
-      const isNotInLifecycleMethod = !inLifeCycleMethod();
+      const isNotInLifecycleMethod = !astUtil.inLifeCycleMethod(context);
       if (isDirectProp && isInClassComponent && isNotInLifecycleMethod) {
         return void 0;
       }
@@ -601,7 +582,7 @@ module.exports = {
             // let {firstname} = props
             const directDestructuring =
               PROPS_REGEX.test(node.init.name) &&
-              (utils.getParentStatelessComponent() || inLifeCycleMethod())
+              (utils.getParentStatelessComponent() || astUtil.inLifeCycleMethod(context))
             ;
 
             if (thisDestructuring) {
@@ -635,7 +616,7 @@ module.exports = {
             name: name,
             allNames: allNames,
             node: (
-              !isDirectProp && !inLifeCycleMethod() ?
+              !isDirectProp && !astUtil.inLifeCycleMethod(context) ?
                 node.parent.property :
                 node.property
             )
@@ -953,7 +934,7 @@ module.exports = {
         const directDestructuring =
           destructuring &&
           PROPS_REGEX.test(node.init.name) &&
-          (utils.getParentStatelessComponent() || inLifeCycleMethod())
+          (utils.getParentStatelessComponent() || astUtil.inLifeCycleMethod(context))
         ;
 
         if (!thisDestructuring && !directDestructuring) {
