+++ lib/util/ast.js
@@ -3,6 +3,8 @@
  */
 'use strict';
 
+const LIFE_CYCLE_METHODS = ['constructor', 'componentWillReceiveProps', 'shouldComponentUpdate', 'componentWillUpdate', 'componentDidUpdate'];
+
 /**
  * Get properties name
  * @param {Object} node - Property.
@@ -59,8 +61,28 @@ function findReturnStatement(node) {
   return false;
 }
 
+/**
+ * Check if we are in a lifecycle method
+ * @return {boolean} true if we are in a lifecycle method, false if not
+ **/
+function inLifeCycleMethod(context) {
+  let scope = context.getScope();
+  while (scope) {
+    if (
+      scope.block && scope.block.parent &&
+      scope.block.parent.key &&
+        LIFE_CYCLE_METHODS.indexOf(scope.block.parent.key.name) >= 0
+    ) {
+      return true;
+    }
+    scope = scope.upper;
+  }
+  return false;
+}
+
 module.exports = {
+  findReturnStatement: findReturnStatement,
   getPropertyName: getPropertyName,
   getComponentProperties: getComponentProperties,
-  findReturnStatement: findReturnStatement
+  inLifeCycleMethod: inLifeCycleMethod
 };
